olympics_back/requirements.txt
flask
python-dotenv
psycopg2-binary
Flask-SQLAlchemy
flask-migrate
python-jose 
flask-cors
flask-jwt-extended